Refactor: Consolidate GitHub integration scripts into github_integration.py (#187)
